backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
passlib[bcrypt]
psycopg2-binary
bcrypt==4.1.2
python-jose[cryptography]==3.3.0